+++ front/src/components/comment/CommentOwner.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CommentOwner = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default CommentOwner;
