+++ front/src/components/common/modules/NotFound.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const NotFound = () => {
+    return (
+        <h1>404 not found</h1>
+    )
+}
+
+export default NotFound;
